+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2024-03-26
+
+### New
+
+- Support for iOS VoiceOver
+- New optional Skip to Content link
+
 ## [2.2.0] - 2022-08-15
 
 ### Changed